# lib/api-zod/src/generated/types/playHistory.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RHF MUSIC API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Song } from './song';

export interface PlayHistory {
  id: number;
  userId: string;
  songId: number;
  playedAt: string;
  song: Song;
}