# lib/api-zod/src/generated/types/playInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RHF MUSIC API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PlayInput {
  userId: string;
  songId: number;
}