# lib/api-zod/src/generated/types/playlistDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RHF MUSIC API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Song } from './song';

export interface PlaylistDetail {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  coverUrl?: string | null;
  userId: string;
  isPublic: boolean;
  songs: Song[];
  createdAt: string;
}