# lib/api-zod/src/generated/types/playlistInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RHF MUSIC API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PlaylistInput {
  name: string;
  description?: string;
  coverUrl?: string;
  userId: string;
  isPublic?: boolean;
}